refactor : 서비스에서 생성자 명시
서비스에서 생성자를 @RequiredArgsConstructor를 사용하지 않고 직접 만들어준다. @Qulifier라는 어노테이션으로 사용할 구현체 레포지토리를 특정한다. 나중에 다른 구현체를 사용하게 되면 다른 구현체 레포지토리의 이름으로 변경하면 된다.
