+++ src/main/java/com/example/quaterback/login/service/JoinService.java
@@ -2,18 +2,23 @@ package com.example.quaterback.login.service;
 
 import com.example.quaterback.exception.DuplicateJoinException;
 import com.example.quaterback.login.entity.UserEntity;
-import com.example.quaterback.login.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
+import com.example.quaterback.login.repository.user.UserRepository;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@RequiredArgsConstructor
 @Service
 public class JoinService {
 
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    public JoinService(@Qualifier("jpaUserRepository") UserRepository userRepository,
+                       BCryptPasswordEncoder bCryptPasswordEncoder) {
+        this.userRepository = userRepository;
+        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
+    }
 
     @Transactional
     public String joinProcess(String username, String password){
